mapper: avoid combining w0 / w1 in access check
As paulb (on Discord) points out w0 and w1 were combined in the wrong
order which made for confusing debug messages.  The underlying mapper
implementation does not need to combine them, so check w0 and w1 flags
separately.

Also |= the UID into the fault, rather than overwriting the fault
with the UID.