mapper: always log proc uid / page uid
The previous code would not log ID mismatches if the proc ID was 0
(due to the way it tested for "was there an ID fault).

But the actual problem appears to be how we handle proc UID 0 :-)

Also swap the proc ID and page ID around to match the text.