mapper: explicitly log if page is not mapped
To assist with debugging.  It's unclear if 0xfff is always used as a sentinel
value for "not mapped", but it is in the cases we have found so far.

no-R/no-W/no-X is likely on both "not mapped" and "notify me when used"
mappings, so we probably cannot rely only no-R/no-W/no-X flags.